Rename 'trias' sandbox to 'ojp'
