+++ application/src/ext-test/java/org/opentripplanner/ext/ojp/trias/OjpMapperTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.trias.trias;
+package org.opentripplanner.ext.ojp.trias;
 
 import static jakarta.xml.bind.Marshaller.JAXB_FORMATTED_OUTPUT;
 import static org.opentripplanner.transit.model._data.TimetableRepositoryForTest.id;
@@ -19,8 +19,8 @@ import java.util.function.Function;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner._support.time.ZoneIds;
 import org.opentripplanner.api.model.transit.DefaultFeedIdMapper;
-import org.opentripplanner.ext.trias.mapping.StopEventResponseMapper;
-import org.opentripplanner.ext.trias.service.CallAtStop;
+import org.opentripplanner.ext.ojp.mapping.StopEventResponseMapper;
+import org.opentripplanner.ext.ojp.service.CallAtStop;
 import org.opentripplanner.model.TripTimeOnDate;
 import org.opentripplanner.transit.model._data.TimetableRepositoryForTest;
 import org.opentripplanner.transit.model.framework.Deduplicator;
