+++ application/src/ext-test/java/org/opentripplanner/ext/ojp/trias/OjpToTriasTransformerTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.trias.trias;
+package org.opentripplanner.ext.ojp.trias;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.framework.io.FileUtils.assertFileEquals;
@@ -15,7 +15,7 @@ import javax.xml.transform.TransformerException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.opentripplanner.ext.trias.mapping.ErrorMapper;
+import org.opentripplanner.ext.ojp.mapping.ErrorMapper;
 import org.opentripplanner.test.support.ResourceLoader;
 
 class OjpToTriasTransformerTest {
