+++ application/src/ext-test/resources/org/opentripplanner/ext/ojp/ojp/stop-event-request-coordinates.xml
