+++ application/src/ext-test/resources/org/opentripplanner/ext/ojp/trias/stop-event-request-coordinates.xml
