+++ application/src/ext/java/org/opentripplanner/ext/ojp/mapping/ServiceDeliveryMapper.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.trias.mapping;
+package org.opentripplanner.ext.ojp.mapping;
 
 import de.vdv.ojp20.siri.ParticipantRefStructure;
 import de.vdv.ojp20.siri.ServiceDelivery;
