+++ application/src/ext/resources/org/opentripplanner/ext/ojp/trias/explorer/api_templates.json
