+++ application/src/main/java/org/opentripplanner/apis/APIEndpoints.java
@@ -25,9 +25,9 @@ import org.opentripplanner.apis.vectortiles.DebugVectorTilesResource;
 import org.opentripplanner.ext.actuator.ActuatorAPI;
 import org.opentripplanner.ext.debugrastertiles.api.resource.DebugRasterTileResource;
 import org.opentripplanner.ext.geocoder.GeocoderResource;
+import org.opentripplanner.ext.ojp.trias.TriasResource;
 import org.opentripplanner.ext.parkAndRideApi.ParkAndRideResource;
 import org.opentripplanner.ext.reportapi.resource.ReportResource;
-import org.opentripplanner.ext.trias.trias.TriasResource;
 import org.opentripplanner.ext.vectortiles.VectorTilesResource;
 import org.opentripplanner.framework.application.OTPFeature;
