+++ application/src/main/java/org/opentripplanner/standalone/config/RouterConfig.java
@@ -10,9 +10,9 @@ import java.io.Serializable;
 import java.util.List;
 import org.opentripplanner.apis.gtfs.GtfsApiParameters;
 import org.opentripplanner.ext.flex.FlexParameters;
+import org.opentripplanner.ext.ojp.config.TriasApiConfig;
+import org.opentripplanner.ext.ojp.parameters.TriasApiParameters;
 import org.opentripplanner.ext.ridehailing.RideHailingServiceParameters;
-import org.opentripplanner.ext.trias.config.TriasApiConfig;
-import org.opentripplanner.ext.trias.parameters.TriasApiParameters;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 import org.opentripplanner.standalone.config.routerconfig.RideHailingServicesConfig;
